https://dagger.dev/tutorial/03-first-command
CONCEPTS

Parameters to an @Inject constructor are the dependencies of the class. Dagger will provide a class’s dependencies to instantiate the class itself. Note that this is recursive: a dependency may have dependencies of its own!
Terminology:
When discussing the relationship between these two types, one might say CommandRouter requests HelloWorldCommand or CommandRouter depends on HelloWorldCommand. Conversely, HelloWorldCommand’s @Inject constructor provides instances of HelloWorldCommand, which are requested by CommandRouter.
Sometimes people say CommandRouter injects HelloWorldCommand in order to emphasize that something else puts a HelloWorldCommand object into CommandRouter, in contrast to CommandRouter itself fetching or creating one. But it’s also common to say that Dagger injects CommandRouter because it instantiates it via its @Inject constructor. These different uses of the word “injects” can get confusing, so we’re using the more explicit terms in this tutorial.