https://dagger.dev/tutorial/05-abstraction-for-output
CONCEPTS

@Provides methods are concrete methods in a module that tell Dagger that when something requests an instance of the type the method returns, it should call that method to get an instance. Like @Inject constructors, they can have parameters: those parameters are their dependencies.
@Provides methods can contain arbitrary code as long as they return an instance of the provided type. They do not need to create a new instance on each invocation.
This highlights an important aspect of Dagger (and dependency injection as a whole): when a type is requested, whether or not a new instance is created to satisfy that request is an implementation detail. Going forward, we’ll use the term “provided” instead of “created”, as that is more accurate for what is happening.