https://dagger.dev/tutorial/06-new-command
CONCEPTS

To start using the LoginCommand in CommandRouter, we’ll replace HelloWorldModule in the @Component annotation with LoginCommandModule. Run the application and try to log in.

This begins to show some of the benefits of using Dagger. With a one line declarative change, we were able to change what Command was received by CommandRouter. CommandRouter had no changes, it just worked. Using this approach, you can write many different versions of your application and reuse code without massive changes.ementation detail. Going forward, we’ll use the term “provided” instead of “created”, as that is more accurate for what is happening.